# Conflicts:
#	README.md

+++ API/src/fr/override/linkit/api/network/cache/HandleableSharedCache.scala
@@ -1,12 +1,12 @@
 package fr.`override`.linkit.api.network.cache
 
-import fr.`override`.linkit.api.packet.channel.CommunicationPacketChannel
+import fr.`override`.linkit.api.packet.fundamental.WrappedPacket
+import fr.`override`.linkit.api.packet.traffic.channel.CommunicationPacketChannel
 import fr.`override`.linkit.api.packet.{Packet, PacketCoordinates}
 import fr.`override`.linkit.api.system.{CloseReason, JustifiedCloseable}
-import fr.`override`.linkit.api.utils.WrappedPacket
 
 abstract class HandleableSharedCache(override val family: String,
-                                     identifier: Int,
+                                     identifier: Long,
                                      channel: CommunicationPacketChannel) extends SharedCache with JustifiedCloseable {
 
     override def close(reason: CloseReason): Unit = channel.close(reason)
